+++ central/telemetry/gatherers/cluster.go
@@ -84,9 +84,6 @@ func (c *ClusterGatherer) Gather(ctx context.Context, pullFromSensors bool) []*d
 		}
 	}
 
-	// Update the maximum node and millicores values for the active clusters.
-	updateMaxima(ctx, clusterList)
-
 	// Get inactive clusters
 	for _, storageCluster := range clusterMap {
 		clusterList = append(clusterList, c.clusterFromDatastores(ctx, storageCluster))
